requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.30.0
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.33.1
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
snowflake-connector-python==3.13.4
